Add a new `yields` keyword.
  JMP <target> YIELDS <mode>
is equivalent to
  JMP <target>
  .mode <mode>
and is used to indicate that the subroutine called leaves the status bits
in a different state than on entry.

Syntactially supported, but not yet hooked up, are
  .entry <mode> <exit_mode>
and
  .remote <address> <mode> <exit_mode>

The former marks an entry point with a non-standard (flags-changing)
calling convention.  The latter tells the disassembler that the given
address is a subroutine with such a calling convention.