Introduce unified syntax for calling conventions.
All instructions representing calls: JSR, JSL, .entry, and .remote, now take an
optional suffix indicating how the subroutine in question returns.

The default calling convention is that the subroutine restores the register
sizes before returning to the caller.

A subroutine that adjusts the flags can indicate this with the suffix
  yields <mode>
whereas a subroutine that never returns (because it manipulates the
return address on the stack, say) can use the suffix
  noreturn

The `disassemble_more` tool honors this flag.  Any calling conventions it learns
about via JSR/JSL/.entry/.remote suffixes are remembered.  If, during disassembly,
a jump to one of these subroutines is found, it is disassembled with the
appropriate syntax.  This changes the behavior of disassembly as well.  That is,
flag state changes and `noreturn` clauses are honored.