Preliminary replacement of FlagState with ExecutionState.
This does not yet track stack values, but tests still pass and the
tools seem to still work.

The Stack class does not yet understand the notion of pushing a
value of unknown size on the stack (e.g., PHA when the M status
bit is unknown), but this will definitely be needed.  Lots of
instruction handling will need to be added in a follow-up as well.

Add a tiny passage to the README file that's been sitting on my
local checkout forever.