Add an Address type, to represent a 65816 address.
The primary benefit of this is we get proper addition and subtraction, always
wrapping within a single bank address.  This fixes a couple of bugs in the
current implementation that didn't fully respect this.

A secondary benefit is that the Location type (used in error reporting) can
differentiate between an `int` (which is a file line number) and an `Address`.

All existing tests pass, but this was a pretty all-encompassing change, so
there might be problems yet to uncover.