Add support for .b and .w assertion suffixes on instructions.
A `.b` or `.w` suffix indicates that the given instruction is expected to operate
with the given data width.  If during assembly, it is found that the code is reached
with the wrong value for a status flag, an error is generated.

For instance, `STZ` and `STZ.b` assemble to the same bytecode, but the latter
asserts that only a single zero byte is being written by the instruction.  If
flow control reaches this instruction with the `m` flag unset, the assertion is
triggered.

A suffix never changes the meaning of code.  For example, given
  ADC #$12
the assembler sees a single-byte immediate operand and will refuse to
assemble it if the current processor state does not allow it.  Adding
a .w suffix, that is:
  ADC.w #$12
does not change the meaning of the code.  The operand size remains incorrect,
and the suffix does not change its data type.

Put another way, a suffix will never change the meaning of an nsasm program.
Adding suffixes may cause a program not to assemble, but otherwise will
have no effect.  Likewise, removing all suffixes from an input file will not
change the output.